Move Themed UI elements into UI folder and update import paths
