+++ codesign/app/(tabs)/profile.tsx
@@ -1,10 +1,10 @@
 import { StyleSheet } from 'react-native';
 
-import { ThemedView } from '@/components/ThemedView';
+import { ThemedView } from '@/components/ui/ThemedView';
 import { EditProfile } from '@/components/profile/EditProfile';
 import { Spacing } from '@/constants/styles/Spacing';
 import { HeaderLayout } from '@/components/shared/HeaderLayout';
-import { TextButton } from '@/components/shared/TextButton';
+import { TextButton } from '@/components/ui/TextButton';
 
 const styles = StyleSheet.create({
   profileContainer: {
