+++ codesign/components/ui/ThemedTextInput.tsx
@@ -6,8 +6,8 @@ import {
 } from 'react-native';
 
 import { useThemeColor } from '@/hooks/useThemeColor';
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
+import { ThemedText } from '@/components/ui/ThemedText';
+import { ThemedView } from '@/components/ui/ThemedView';
 import { Spacing } from '@/constants/styles/Spacing';
 import { Typography } from '@/constants/styles/Typography';
